알림창 수정 (#90)
* #88 - feat: 로그인/회원가입 알림창 수정

* #88 -feat: 알림창 수정
