+++ src/components/Calendar/History/Comparison/ComparisonPrevMonth.tsx
@@ -1,5 +1,5 @@
 import { useAppSelector } from '@/hooks/store';
-import { HistoryProps } from '@/pages/Calendar';
+import { HistoryProps } from '@/components/Layout/Route';
 
 function ComparisonPrevMonth({ history }: HistoryProps) {
   const calendar = useAppSelector((state) => state.calendar);
