+++ src/components/Calendar/History/HistoryDetail.tsx
@@ -1,7 +1,7 @@
-import { HistoryProps } from '@/pages/Calendar';
 import * as S from '@/styles/Calendar/History/HistoryDetail.style';
 import HistoryHeader from '@/components/Calendar/Header/HistoryHeader';
 import HistoryMain from '@/components/Calendar/Main/HistoryMain';
+import { HistoryProps } from '@/components/Layout/Route';
 
 export interface UpdateModalProps {
   updateModal: boolean;
