+++ src/components/Calendar/History/Modal/AddHistoryForm.tsx
@@ -8,6 +8,8 @@ import { AddModalProps } from '@/components/Calendar/History/CalendarHistory';
 import { baseAPI } from '@/api/customAxios';
 import { onToggle } from '@/store/historySlice';
 import Alert from '@/components/Alert/Alert';
+import { ModalWrapper } from '@/styles/Global/modal.style';
+import DateModal from '@/components/Calendar/Modal/DateModal';
 
 export interface FormState {
   id?: number;
@@ -31,8 +33,8 @@ function AddHistoryForm({ addModal, setAddModal }: AddModalProps) {
   const [formState, setFormState] = useState<FormState>({});
   const [alert, setAlert] = useState(false);
 
-  const onSubmit = async (event: React.FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
+  const onSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
+    e.preventDefault();
 
     const reqbody = {
       year: calendar.year,
@@ -47,88 +49,99 @@ function AddHistoryForm({ addModal, setAddModal }: AddModalProps) {
 
     try {
       await baseAPI.post('/historys', reqbody);
-      setAddModal(false);
       dispatch(onToggle(true));
-    } catch (err) {
-      console.log(err);
-    }
+      setAddModal(false);
+      setAlert(true);
+    } catch {}
   };
 
   return (
-    <S.ItemContainer onSubmit={onSubmit}>
-      <S.ItemBox>
-        <p className="itemTitle">분류</p>
-        <p className="itemContent">
-          <SelectClass formState={formState} setFormState={setFormState} />
-        </p>
-      </S.ItemBox>
-      <S.ItemBox>
-        <p className="itemTitle">카테고리</p>
-        <p className="itemContent">
-          {formState.value === '수입' ? (
-            <IncomeCategory formState={formState} setFormState={setFormState} />
-          ) : (
-            <ExpensesCategory
-              formState={formState}
-              setFormState={setFormState}
+    <>
+      <ModalWrapper isOpen={addModal} onClick={() => setAddModal(false)} />
+
+      <S.ModalContainer modal={addModal}>
+        <DateModal />
+
+        <S.ItemContainer onSubmit={onSubmit}>
+          <S.ItemBox>
+            <p className="itemTitle">분류</p>
+            <p className="itemContent">
+              <SelectClass formState={formState} setFormState={setFormState} />
+            </p>
+          </S.ItemBox>
+          <S.ItemBox>
+            <p className="itemTitle">카테고리</p>
+            <p className="itemContent">
+              {formState.value === '수입' ? (
+                <IncomeCategory
+                  formState={formState}
+                  setFormState={setFormState}
+                />
+              ) : (
+                <ExpensesCategory
+                  formState={formState}
+                  setFormState={setFormState}
+                />
+              )}
+            </p>
+          </S.ItemBox>
+          <S.ItemBox>
+            <label htmlFor="amount" className="itemTitle">
+              금액
+            </label>
+            <input
+              type={'number'}
+              id="amount"
+              className="underline itemContent"
+              value={formState.amount || ''}
+              onChange={(e) =>
+                setFormState({
+                  ...formState,
+                  amount: e.target.valueAsNumber
+                })
+              }
+              required
             />
-          )}
-        </p>
-      </S.ItemBox>
-      <S.ItemBox>
-        <label htmlFor="amount" className="itemTitle">
-          금액
-        </label>
-        <input
-          type={'number'}
-          id="amount"
-          className="underline itemContent"
-          value={formState.amount || ''}
-          onChange={(e) =>
-            setFormState({
-              ...formState,
-              amount: e.target.valueAsNumber
-            })
-          }
-          required
-        />
-      </S.ItemBox>
-      <S.ItemBox>
-        <label htmlFor="content" className="itemTitle">
-          내용
-        </label>
-        <input
-          type={'text'}
-          id="content"
-          className="underline itemContent"
-          value={formState.content || ''}
-          onChange={(e) =>
-            setFormState({ ...formState, content: e.target.value })
-          }
-          required
-        />
-      </S.ItemBox>
-      <S.ButtonBox>
-        <div className="option">
-          <button type="submit" className="blue">
-            추가
-          </button>
-          <Alert
-            message="내역이 추가되었습니다."
-            trueText="확인"
-            alertModal={alert}
-            setAlertModal={setAlert}
-          />
-          <button
-            type="reset"
-            className="red"
-            onClick={() => setAddModal(false)}
-          >
-            취소
-          </button>
-        </div>
-      </S.ButtonBox>
-    </S.ItemContainer>
+          </S.ItemBox>
+          <S.ItemBox>
+            <label htmlFor="content" className="itemTitle">
+              내용
+            </label>
+            <input
+              type={'text'}
+              id="content"
+              className="underline itemContent"
+              value={formState.content || ''}
+              onChange={(e) =>
+                setFormState({ ...formState, content: e.target.value })
+              }
+              required
+            />
+          </S.ItemBox>
+          <S.ButtonBox>
+            <div className="option">
+              <button type="submit" className="blue">
+                추가
+              </button>
+              <button
+                type="reset"
+                className="red"
+                onClick={() => setAddModal(false)}
+              >
+                취소
+              </button>
+            </div>
+          </S.ButtonBox>
+        </S.ItemContainer>
+      </S.ModalContainer>
+
+      <Alert
+        message="내역이 추가되었습니다."
+        trueText="확인"
+        alertModal={alert}
+        setAlertModal={setAlert}
+      />
+    </>
   );
 }
