+++ src/components/Calendar/Main/CalendarMain.tsx
@@ -1,7 +1,7 @@
 import * as S from '@/styles/Calendar/Main/CalendarMain.style';
 import CalendarDay from '@/components/Calendar/Main/CalendarDay';
 import CalendarDate from '@/components/Calendar/Main/CalendarDate';
-import { HistoryProps } from '@/pages/Calendar';
+import { HistoryProps } from '@/components/Layout/Route';
 
 function Main({ history }: HistoryProps) {
   return (
