+++ src/pages/Chart.tsx
@@ -1,50 +1,11 @@
 import BarChart from '@/components/Chart/BarChart';
 import ExpensesPieChart from '@/components/Chart/ExpensesPieChart';
 import IncomePieChart from '@/components/Chart/IncomePieChart';
-import { useEffect, useState } from 'react';
 import Header from '@/components/Calendar/Header/CalendarHeader';
-import { baseAPI } from '@/api/customAxios';
 import * as S from '@/styles/Chart/Chart.style';
+import { HistoryProps } from '@/components/Layout/Route';
 
-type History = {
-  id: number;
-  userId: number;
-  year: number;
-  month: number;
-  date: number;
-  value: string;
-  category: string;
-  amount: number;
-  content: string;
-}[];
-
-export interface ChartHistoryProps {
-  history: History;
-  setHistory?: React.Dispatch<React.SetStateAction<History>>;
-}
-
-function Chart() {
-  const [history, setHistory] = useState<History>([]);
-
-  const getHistory = async () => {
-    try {
-      const res = await baseAPI.get('/historys');
-      setHistory(
-        res.data.filter(
-          (el: any) =>
-            el.userId ===
-            (localStorage.user ? JSON.parse(localStorage.user).userId : 0)
-        )
-      );
-    } catch (err) {
-      console.log(err);
-    }
-  };
-
-  useEffect(() => {
-    getHistory();
-  }, []);
-
+function Chart({ history }: HistoryProps) {
   return (
     <S.ChartWrapper>
       <S.ChartContainer>
