+++ src/styles/Sign/Signup.style.ts
@@ -19,6 +19,7 @@ export const SignUpFormContent = styled.form`
 
   .input_box {
     display: flex;
+    align-items: center;
     width: 100%;
     margin-bottom: 15px;
     font-size: 9px;
@@ -41,25 +42,5 @@ export const SignUpFormContent = styled.form`
     color: red;
     font-size: 8px;
     margin-top: -15px;
-  }
-`;
-
-export const SignUpButton = styled.button`
-  width: 100%;
-  margin-top: 9px;
-  padding: 0.5rem 1rem;
-  border-radius: 4px;
-  border: none;
-  background-color: ${(props) => props.theme.main};
-  color: ${(props) => props.theme.white};
-  font-size: 1rem;
-  cursor: pointer;
-
-  @media ${(props) => props.theme.desktop} {
-    transition: background-color 0.5s;
-
-    &:hover {
-      background-color: ${(props) => props.theme.hover};
-    }
   }
 `;
